proofreadpage_tests: Require bs4 for setUpClass
The class IndexPageTestCase earlier checked for bs4 as a
class decorator. But that check doesn't seem to work and is
a known bug in node. We add the required_module() decorator
to the setUpClass also so that it is not run.

The related issue on nose's github repository can be found at
https://github.com/nose-devs/nose/issues/946

Bug: T129965
Change-Id: Iec2f706aa832f7bd20e56fbc3427e8f315499591