pywikibot: Store ImportError in imported variable
If a lazy ImportError is to be handled, the ImportError
needs to be stored in a vairable and then used when required.
Earlier, the imported vairable was set to `None` and the
import error which was caught with `except` was used. But
this import error may not exist at a later time as python
guarantees it's existence only inside the `except` statement.
Hence, this is changed so that the ImportError is saved in the
variable that was being imported and checked later using
`isinstance(<variable>, ImportError)`.

Bug: T134337
Change-Id: I80f82e7f0674bfca70688f28640620dacac1d80b